Q.50/requirements.txt
# Requirements for Telecom Big Data Architecture
boto3==1.34.0
pyflink==1.16.0
pyspark==3.4.0
delta-spark==2.4.0
sagemaker==2.226.0
pandas==2.2.0